requirements.txt
pandas==2.2.2
python-dateutil==2.9.0
requests==2.32.3
tenacity==9.0.0
python-dotenv==1.0.1
gspread==6.1.2
gspread-dataframe==3.3.1
numpy==2.0.1
